requirements.txt
numpy
pandas
scikit-learn
pyyaml
pytest
joblib